refactor debuginfoproviders and make them available for ICSharpCode.Decompiler
