+++ ICSharpCode.Decompiler/DebugInfo/PortableDebugInfoProvider.cs
@@ -22,9 +22,9 @@ using System.Reflection.Metadata;
 using ICSharpCode.Decompiler.DebugInfo;
 using ICSharpCode.Decompiler.Metadata;
 
-namespace ICSharpCode.ILSpy.DebugInfo
+namespace ICSharpCode.Decompiler.PdbProvider
 {
-	class PortableDebugInfoProvider : IDebugInfoProvider
+	public class PortableDebugInfoProvider : IDebugInfoProvider
 	{
 		string pdbFileName;
 		MetadataReaderProvider provider;
