+++ ICSharpCode.Decompiler/Metadata/PEFile.cs
@@ -23,6 +23,9 @@ using System.IO;
 using System.Linq;
 using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
+using System.Runtime.InteropServices;
+using ICSharpCode.Decompiler.DebugInfo;
+using ICSharpCode.Decompiler.PdbProvider;
 using ICSharpCode.Decompiler.TypeSystem;
 using ICSharpCode.Decompiler.Util;
 
@@ -111,6 +114,70 @@ namespace ICSharpCode.Decompiler.Metadata
 		public void Dispose()
 		{
 			Reader.Dispose();
+		}
+
+		public IDebugInfoProvider LoadSymbols()
+		{
+			try {
+				var reader = Reader;
+				// try to open portable pdb file/embedded pdb info:
+				if (TryOpenPortablePdb(out var provider, out var pdbFileName)) {
+					return new PortableDebugInfoProvider(pdbFileName, provider);
+				} else {
+					// search for pdb in same directory as dll
+					string pdbDirectory = Path.GetDirectoryName(this.FileName);
+					pdbFileName = Path.Combine(pdbDirectory, Path.GetFileNameWithoutExtension(this.FileName) + ".pdb");
+					if (File.Exists(pdbFileName)) {
+						return new MonoCecilDebugInfoProvider(this, pdbFileName);
+					}
+
+					// TODO: use symbol cache, get symbols from microsoft
+				}
+			} catch (Exception ex) when (ex is BadImageFormatException || ex is COMException) {
+				// Ignore PDB load errors
+			}
+			return null;
+		}
+
+		const string LegacyPDBPrefix = "Microsoft C/C++ MSF 7.00";
+		byte[] buffer = new byte[LegacyPDBPrefix.Length];
+
+		bool TryOpenPortablePdb(out MetadataReaderProvider provider, out string pdbFileName)
+		{
+			provider = null;
+			pdbFileName = null;
+			var reader = Reader;
+			foreach (var entry in reader.ReadDebugDirectory()) {
+				if (entry.IsPortableCodeView) {
+					return reader.TryOpenAssociatedPortablePdb(this.FileName, OpenStream, out provider, out pdbFileName);
+				}
+				if (entry.Type == DebugDirectoryEntryType.CodeView) {
+					string pdbDirectory = Path.GetDirectoryName(this.FileName);
+					pdbFileName = Path.Combine(pdbDirectory, Path.GetFileNameWithoutExtension(this.FileName) + ".pdb");
+					if (File.Exists(pdbFileName)) {
+						Stream stream = OpenStream(pdbFileName);
+						if (stream.Read(buffer, 0, buffer.Length) == LegacyPDBPrefix.Length
+							&& System.Text.Encoding.ASCII.GetString(buffer) == LegacyPDBPrefix) {
+							return false;
+						}
+						stream.Position = 0;
+						provider = MetadataReaderProvider.FromPortablePdbStream(stream);
+						return true;
+					}
+				}
+			}
+			return false;
+		}
+
+		Stream OpenStream(string fileName)
+		{
+			if (!File.Exists(fileName))
+				return null;
+			var memory = new MemoryStream();
+			using (var stream = File.OpenRead(fileName))
+				stream.CopyTo(memory);
+			memory.Position = 0;
+			return memory;
 		}
 
 		Dictionary<TopLevelTypeName, TypeDefinitionHandle> typeLookup;
