+++ ILSpy/LoadedAssembly.cs
@@ -23,15 +23,12 @@ using System.IO;
 using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 using ICSharpCode.Decompiler.DebugInfo;
 using ICSharpCode.Decompiler.Metadata;
-using ICSharpCode.Decompiler.PdbProvider.Cecil;
 using ICSharpCode.Decompiler.TypeSystem;
 using ICSharpCode.Decompiler.TypeSystem.Implementation;
-using ICSharpCode.ILSpy.DebugInfo;
 using ICSharpCode.ILSpy.Options;
 
 namespace ICSharpCode.ILSpy
@@ -163,7 +160,7 @@ namespace ICSharpCode.ILSpy
 
 			if (DecompilerSettingsPanel.CurrentDecompilerSettings.UseDebugSymbols) {
 				try {
-					LoadSymbols(module);
+					debugInfoProvider = module.LoadSymbols();
 				} catch (IOException) {
 				} catch (UnauthorizedAccessException) {
 				} catch (InvalidOperationException) {
@@ -176,69 +173,7 @@ namespace ICSharpCode.ILSpy
 			return module;
 		}
 		
-		void LoadSymbols(PEFile module)
-		{
-			try {
-				var reader = module.Reader;
-				// try to open portable pdb file/embedded pdb info:
-				if (TryOpenPortablePdb(module, out var provider, out var pdbFileName)) {
-					debugInfoProvider = new PortableDebugInfoProvider(pdbFileName, provider);
-				} else {
-					// search for pdb in same directory as dll
-					string pdbDirectory = Path.GetDirectoryName(fileName);
-					pdbFileName = Path.Combine(pdbDirectory, Path.GetFileNameWithoutExtension(fileName) + ".pdb");
-					if (File.Exists(pdbFileName)) {
-						debugInfoProvider = new MonoCecilDebugInfoProvider(module, pdbFileName);
-						return;
-					}
 
-					// TODO: use symbol cache, get symbols from microsoft
-				}
-			} catch (Exception ex) when (ex is BadImageFormatException || ex is COMException) {
-				// Ignore PDB load errors
-			}
-		}
-
-		const string LegacyPDBPrefix = "Microsoft C/C++ MSF 7.00";
-		byte[] buffer = new byte[LegacyPDBPrefix.Length];
-
-		bool TryOpenPortablePdb(PEFile module, out MetadataReaderProvider provider, out string pdbFileName)
-		{
-			provider = null;
-			pdbFileName = null;
-			var reader = module.Reader;
-			foreach (var entry in reader.ReadDebugDirectory()) {
-				if (entry.IsPortableCodeView) {
-					return reader.TryOpenAssociatedPortablePdb(fileName, OpenStream, out provider, out pdbFileName);
-				}
-				if (entry.Type == DebugDirectoryEntryType.CodeView) {
-					string pdbDirectory = Path.GetDirectoryName(fileName);
-					pdbFileName = Path.Combine(pdbDirectory, Path.GetFileNameWithoutExtension(fileName) + ".pdb");
-					if (File.Exists(pdbFileName)) {
-						Stream stream = OpenStream(pdbFileName);
-						if (stream.Read(buffer, 0, buffer.Length) == LegacyPDBPrefix.Length
-							&& System.Text.Encoding.ASCII.GetString(buffer) == LegacyPDBPrefix) {
-							return false;
-						}
-						stream.Position = 0;
-						provider = MetadataReaderProvider.FromPortablePdbStream(stream);
-						return true;
-					}
-				}
-			}
-			return false;
-		}
-
-		Stream OpenStream(string fileName)
-		{
-			if (!File.Exists(fileName))
-				return null;
-			var memory = new MemoryStream();
-			using (var stream = File.OpenRead(fileName))
-				stream.CopyTo(memory);
-			memory.Position = 0;
-			return memory;
-		}
 
 		[ThreadStatic]
 		static int assemblyLoadDisableCount;
